Rooms and Secret Doors
I first rearranged the rooms so that the layout would move east and
west as oppose to the original north and south. That way when the
layout of the house is flipped horizontally upon going to the other
dimension, the change in direction is actually noticable. I also
finished creating the parallel dimension rooms and door by changing the
printed names of all of the alternate deminsion rooms to match their
original counterparts.

I also added in the extension Secret Doors by Andrew Owen. This took a
little while to make work because the source code for the extension
comes with 2 bugs in which 'when when' is said, messing with the code.
Word has already been spread to several classmates who wish to use this
extension but it is good to keep as a note.