Rooms, Items, Descriptions, NPC
I finished the room layout with identical printed names for the
alternate dimension rooms. I added in new items for future puzzles such
as a suitcase to be filled with clothes and other toliteries you'll
need to get ready. I am also setting up items such as a clock and
calender for you to prove that you are in another dimension (another
puzzle). An NPC has also been created with a set up to prepare for
dialouge and responses.