Begining of Game, Clothes, Scenery, NPC
I have set the stage for the begining of the game before the main plot
occurs. I have programmed the player to wake up in bed and in dirty
clothes. I have also created a suitcase with clean clothes inside for
the player to change into to prepare for the arrival of Ellie, the
player's girlfriend.

I have also changed some of the scenery and such. I added a dresser
drawer as an openable container to compliment the dresser which is a
scenery supporter. This will perhaps be the first place the player will
look for clean clothes, however I still need to let the player know
they haven't unpacked yet and that clothes will be else where in the
suitcase.

I also added a front door to the living room. It won't actually open
but will act as a logical addition for which you house will have and
Ellie will enter through. Ellie is also the NPC of your real
girlfriend. Work needs to be done regarding her responses and movement
but code has been set up to prepare such.