Bed, Dresser Drawer, Clothes, Elle
I just did a few tweaks in the code for clarities sake regarding the
first basic challenge of the game, which is to get dressed before Elle
arrives. As this response is triggered after leaving the bed, I coded
the game to understand getting in and out as climbing as prior to this
the game responded only to getting on and off. Just a simple addition
of synonyms like the reading told us to make sure to do. I also wrote
the code so that if you try to look in the dresser drawer at first for
clothes, the game will tell you clean clothes are located elsewhere in
the suitcase.  I still need to add to Elle's entrance so that you
discover the door. I also would like to fix it so that when you examine
yourself you see the clothes you're wearing but I don't have a fix for
that at the moment so I have simply made note of it for later.