Continuity Errors and Other World
I added in an xyzzy command for later use. I'll make it something more
interesting later but for now I just got the code in.

I fixed the code so that the game understands 'get out of bed as 'get
off of bed.' Just a samll touch but the first fix didn't quite do the
trick.

I also fixed the couch in the real world and made a copy for the other
world. I also put in the mail in both worlds that will act as a clue to
the reality as they have very different meanings. Move rug is also now
understood as inform had trouble understanding the verb even though it
understands move in game. I also had to change the clean clothes
response because after I put them on in a weird order it didn't display
the next message. Kapri has told me sometimes using 'after' conditions
doesn't work and that 'instead of' usually does.

I also fixed a lot of continuity errors, mainly the doors in the game.
The mirror (portal) is now an open door so it doesn't say 'first
opening mirror' when you go through. The hidden door in the other world
(hidden door1) is also already open to simulate how you left the real
one in the first world. The front door to the house has also been
modified to only display the update message the first time you open it.
I accidently discovered it would continue to display it multiple times.

As already mentioned I did more to further the flow of the plot of the
game. Mail is now an object that will clue you in to the reality of
multiple worlds. I also programmed Elle to follow you through the
rooms. Hopefully these commands will hold up even if you approach the
challenges in a different order.