Ending
I created an end to the game and have made it winnable. I also made a
lot of conditions and responses regarding both of my NPCs. I also
finalized many of the items and their roles such as the beverages, and
clues: Mail, Clock, Calender. Using two NPCs was difficult because
because I needed to try an account for the player doing different
things in different orders. However, i still needed the game to steer
the player down the right path for the plot.

After bringing together all the clues with Elle you must escape back to
your own world. This part is mainly for plot and story but still
requires action by the player.